Big commit, lol

+++ src/main/java/com/epam/jwd_critics/entity/AgeRestriction.java
@@ -1,23 +1,24 @@
 package com.epam.jwd_critics.entity;
 
+
 import com.epam.jwd_critics.exception.UnknownEntityException;
 
 import java.util.Arrays;
 
 public enum AgeRestriction implements BaseEntity{
-    G(1L),
-    PG(2L),
-    PG_13(3L),
-    R(4L),
-    NC_17(5L);
-    private final Long id;
+    G(1),
+    PG(2),
+    PG_13(3),
+    R(4),
+    NC_17(5);
+    private final Integer id;
 
-    AgeRestriction(Long id) {
+    AgeRestriction(Integer id) {
         this.id = id;
     }
 
     @Override
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
