+++ src/main/java/com/epam/jwd_critics/entity/Celebrity.java
@@ -3,7 +3,7 @@ package com.epam.jwd_critics.entity;
 public class Celebrity extends AbstractBaseEntity{
     private String name;
 
-    Celebrity(Long id, String name) {
+    Celebrity(Integer id, String name) {
         super(id);
         this.name = name;
     }
